requirements.txt
# Healthcare Information Guide Agent — Dependencies
langchain==0.2.16
langchain-groq==0.1.9
langchain-community==0.2.16
pandas==2.2.2
datasets==2.21.0
requests==2.32.3
python-dotenv==1.0.1
tqdm==4.66.5
